Fix the stats page display
This commit fixes the display of the Package Stats that was broken
with the recent changes.

I also reduce the number of calls to the database by storing the
results of the queries maintain_select and top_owners_select in
memory. This way they are ran once instead of twice.